python/requirements.txt
# Django core
Django>=4.2,<5.0
django-environ>=0.11.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
django-allauth>=0.60.0

# Internationalization
django-modeltranslation>=0.18.0
django-rosetta>=0.10.0

# Admin & UI
django-jazzmin>=2.6.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10

# API
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Task scheduling
django-apscheduler>=0.6.2
APScheduler>=3.10.0

# Integrations
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0
requests>=2.31.0
paramiko>=3.4.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development
django-debug-toolbar>=4.2.0
